requirements.txt
numpy >= 2.4.4
matplotlib >= 3.10.9
pandas >= 3.0.2
PyOpenGL >= 3.1.10
wxPython >= 4.2.5
pyrr >= 0.10.3
imageio >= 2.37.3
pyserial >= 3.5
h5py >= 3.16.0
sympy >= 1.14.0